Exit `docc` preview server upon Ctrl+C
Resolves a bug where the preview-documentation server would fail to
start upon subsequence runs because the interrupt sent to the parent
SwiftPM process (via ctrl+c) was not passed along to the `docc` process.

Since Foundation does not expose API to create a process in the same
group as the parent, the solution is to monitor for SIGKILL and SIGINT
and pass along these signals to the child `docc` process.