Adopt new SwiftPM argument parsing API
The SwiftPM plugin API has been updated so that options like `--target`
and `--product` are passed directly to the plugin instead of to
the package manager.

This means that instead of calling the plugin something like:

    swift package --target SwiftMarkdown generate-documentation

you would now do:

    swift package generate-documentation --target SwiftMarkdown

This allows us to be a little smarter about choosing which products to build.

For example, now, by default, when a user invokes the basic

    swift package generate-documentation

we will also generate documentation for dependencies. Similarly, this allows
clients to specify products of dependencies which makes doing something like

    swift package preview-documetnation --product ArgumentParser

possible for packages that _import_ ArgumentParser.

Adopt new SwiftPM argument parsing API

The SwiftPM plugin API has been updated so that options like `--target`
and `--product` are passed directly to the plugin instead of to
the package manager.

This means that instead of calling the plugin something like:

    swift package --target SwiftMarkdown generate-documentation

you would now do:

    swift package generate-documentation --target SwiftMarkdown

This allows us to be a little smarter about choosing which products to build.

For example, now, by default, when a user invokes the basic

    swift package generate-documentation

we will also generate documentation for dependencies. Similarly, this allows
clients to specify products of dependencies which makes doing something like

    swift package preview-documentation --product ArgumentParser

possible for packages that _import_ ArgumentParser.