Preview: remove use of `signal` on Windows (#26)
Windows does not support signal(2).  Instead, use `TerminateProcess` to
terminate the process.  We will not properly handle the termination of
child processes for the preview process, which should be handled by the
construction of a job object that all child processes are attached to
such that when the parent process is terminated, the children will be
properly reaped.  This is required to enable building the plugin on
Windows which is a dependency for swift-syntax.